feat(profile): let students see their profile exactly as others do
The right rail card said "this is how your profile appears to instructors and
recruiters", which was a claim the page could not back up: it showed a name, a
headline and a location, not the profile. Students asked to actually see it.

PublicProfileView is the read-only rendering of a profile, extracted from the
admin page, and both surfaces now render it:

  app/admin/profile/[id]  an admin inspecting a student
  app/profile/preview     a student checking their own public view

Sharing one component rather than building a second lookalike is the point. The
ask was to see "exactly" what other people see, and two components that merely
resemble each other answer that on the day they are written and then drift. This
way the answer stays true because there is only one render.

The preview is read-only by construction rather than by a flag: no save handler
is threaded in, and PublicProfileView passes no edit callbacks down, so nothing
on that route can mutate the profile. It also reads the raw API profile with no
localStorage merge, unlike the edit page, because showing an unsaved local draft
in a "what others see" view would be a lie.

The completion stat row stays admin-only. It is internal chrome about a student,
not something a viewer of the profile sees.
